moved rotaxis() tests to tests for mdamath
