+++ testsuite/MDAnalysisTests/utils/test_transformations.py
@@ -33,8 +33,9 @@ from numpy.testing import (
 
 from MDAnalysis.lib import transformations as t
 
-from unittest import TestCase
-
+# ------------------------------------------------------------
+# DEPRECATION: Remove this file in 3.0 when lib.transformations is removed
+# ------------------------------------------------------------
 
 """
 Testing transformations is weird because there are 2 versions of many of
@@ -48,6 +49,9 @@ This should ensure that both versions work and are covered!
 .. versionchanged:: 1.0.0
    test_transformations_old_module was removed as core/transformations.py is
    gone
+
+.. deprecated:: 2.10.0
+   will be removed in 3.0.
 """
 
 # tolerance for tests
@@ -59,8 +63,8 @@ _ATOL = 1e-06
 # versions from transformations._transformations seem to always shadow
 # the Python ones from transformations.transformations.
 #
-# As a hack, I replaced t._py_* with t.transformations.* so that the
-# tests run but they just test the compiled code twice.
+# As a hack, I replaced t._py_* with t.transformations.* so that the tests run
+# but they just test the compiled code twice. (See PR GH-5062).
 
 
 @pytest.mark.parametrize(
@@ -420,23 +424,10 @@ def test_superimposition_matrix(f):
     M = f(v0, v1)
     assert_allclose(v1, np.dot(M, v0), atol=_ATOL)
 
-    # scaling kwarg not supported in transformations any more
-    #
-    # S = t.scale_matrix(0.45)
-    # T = t.translation_matrix(np.array([0.2, 0.2, 0.2]) - 0.5)
-    # M = t.concatenate_matrices(T, R, S)
-    # v1 = np.dot(M, v0)
-    # v0[:3] += np.sin(np.linspace(0.0, 1e-9, 300)).reshape(3, -1)
-    # M = f(v0, v1, scaling=True)
-    # assert_allclose(v1, np.dot(M, v0), atol=_ATOL)
-
-    # M = f(v0, v1, scaling=True, usesvd=False)
-    # assert_allclose(v1, np.dot(M, v0), atol=_ATOL)
-    #
-    # v = np.empty((4, 100, 3), dtype=np.float64)
-    # v[:, :, 0] = v0
-    # M = f(v0, v1, scaling=True, usesvd=False)
-    # assert_allclose(v1, np.dot(M, v[:, :, 0]), atol=_ATOL)
+    # We used to have tests here that also tested the `scaling=True` kwarg of
+    # transformations.superimposition_matrix(). This kwarg is no longer
+    # supported in the transformations package. We don't need it for MDAnalysis
+    # so we just removed the tests. (See PR GH-5062)
 
 
 @pytest.mark.parametrize(
@@ -841,31 +832,9 @@ class TestArcBall(object):
         assert_allclose(np.sum(R), 0.2055924)
 
 
-# rotaxis was an MDA addition
-
-
-def test_rotaxis_equal_vectors():
-    a = np.arange(3)
-    x = t.rotaxis(a, a)
-    assert_array_equal(x, [1, 0, 0])
-
-
-def test_rotaxis_different_vectors():
-    # use random coordinate system
-    e = np.eye(3)
-    r = np.array(
-        [
-            [0.69884766, 0.59804425, -0.39237102],
-            [0.18784672, 0.37585347, 0.90744023],
-            [0.69016342, -0.7078681, 0.15032367],
-        ]
-    )
-    re = np.dot(r, e)
-
-    for i, j, l in permutations(range(3)):
-        x = t.rotaxis(re[i], re[j])
-        # use abs since direction doesn't matter
-        assert_almost_equal(np.abs(np.dot(x, re[l])), 1)
+# rotaxis() was an MDA addition. It was migrated to lib.mdamath in 2.10.0 (and
+# its functionality is tested there). A stub with deprecation warning was left
+# behin in lib.transformations. All of it will be removed in 3.0.0.
 
 
 def test_rotaxis_deprecation():
